Create seperate 'ReadOnlyMask256'
Instead of having a mutable flag on the 'Mask256', this way
its nicer to use as if you have a 'Mask256' you know you can
mutate it and if you have a 'ReadOnlyMask256' you cannot.

Also its a tiny perf boost as the struct is a bit smaller and now
nicely aligns to 256 bit.