Replaced avx2 implementation of 'Find' with a simple loop.
Reason is that after benchmarking i found that in many cases its
actually slower.
If the element to find is less then 4 slots away for ideal position
its actually faster to just loop instead of using intrinsics.
If its 4 or more slots away the intrinsics path is faster but most
often the item is not that far from ideal.